+++ js/util.js
@@ -5,14 +5,12 @@ const getRandomIntInclusive = (min, max) => {
     return Math.floor(Math.random() * (max - min + 1)) + min;
   }
 }
-getRandomIntInclusive(1, 15);
 
 const getRandomFloatInclusive = (min, max, digits) => {
   if (min >= 0 && min <= max && digits >= 0) {
     return parseFloat((Math.random() * (max - min) + min).toFixed(digits));
   }
 }
-getRandomFloatInclusive(2.15, 27.445, 5);
 
-export {getRandomIntInclusive};
-export {getRandomFloatInclusive};
+export {getRandomIntInclusive, getRandomFloatInclusive};
+
